New from Linux

+++ bash_init/bash_init.sh
@@ -2,10 +2,10 @@
 # version: 1.0
 
 #!/bin/bash
-mv $HOME/.bash_aliases $HOME/.bash_aliases.orig
-mv $HOME/.bash_profile $HOME/.bash_profile.orig
-mv $HOME/.bash_progs $HOME/.bash_progs.orig
-mv $HOME/.bashrc $HOME/.bashrc.orig
+if [[ -e $HOME/.bash_aliases ]]; then mv $HOME/.bash_aliases $HOME/.bash_aliases.orig; fi
+if [[ -e $HOME/.bash_profile ]]; then mv $HOME/.bash_profile $HOME/.bash_profile.orig; fi
+if [[ -e $HOME/.bash_progs ]]; then mv $HOME/.bash_progs $HOME/.bash_progs.orig; fi
+if [[ -e $HOME/.bashrc ]]; then mv $HOME/.bashrc $HOME/.bashrc.orig; fi
 
 ln -s $HOME/git/lnx_scripts/bash_aliases/.bash_aliases $HOME/.bash_aliases
 ln -s $HOME/.bashrc $HOME/.bash_profile
